deployment/eventbusandruleexample/requirements.txt
aws-cdk-lib==2.36.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0